Stack implementation file added

+++ libft/ft_atoi.c
@@ -6,11 +6,13 @@
 /*   By: vpaliash <vpaliash@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/03 14:14:07 by vpaliash          #+#    #+#             */
-/*   Updated: 2024/12/16 17:21:20 by vpaliash         ###   ########.fr       */
+/*   Updated: 2025/03/20 13:44:44 by vpaliash         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_atoi(const char *str)
+#include "libft.h"
+
+size_t	ft_atoi(const char *str)
 {
 	int	result;
 	int	sign;
